updated images (except portfolio and products cos I've got to work out what I'm doing there
