+++ src/img/favicons/favicon-96.png
