+++ src/js/data/services.js
@@ -1,0 +1,52 @@
+const services = [
+    {
+        name: 'Website Development',
+        slug: 'web-dev',
+        description: 'We specialise in creating custom websites for you and your business. We work with each client individually to find and build the best solution for your needs.',
+        price: '',
+    },
+    {
+        name: 'Email Hosting',
+        slug: 'email',
+        description: 'Get a custom email for your business (me@mybusiness.com.au). Hosting is provided through either Google G Suite or Microsoft Office 365 and includes online storage. Need Microsoft Office as well? We can bundle it in.',
+        price: '$6/month',
+    },
+    {
+        name: 'Ongoing Support and Maintenance',
+        slug: 'aftersale',
+        description: "Websites shouldn't just be 'set and forget'. We're always there for any questions that you have (except if it's 3am). We also offer customised maintenance plans to help manage your website, so you don't have to.",
+        price: '$15/month',
+    },
+    {
+        name: 'Website Hosting',
+        slug: 'web-host',
+        description: "We can help you get your website online (and stay there). Don't worry about the technical aspects, let us do that for you.",
+        price: '$6/month',
+    },
+    {
+        name: 'Content Writing and Editing',
+        slug: 'content',
+        description: 'Having trouble stringing your website content together or just need someone to look over it? We can help read over your content, or research and write your website content for you.',
+        price: '',
+    },
+    {
+        name: 'Search Engine Optimisation',
+        slug: 'seo',
+        description: 'Search Engine Optimisation (SEO) helps you to appear in Google search results. Let us help you with your top-level SEO needs.',
+        price: '',
+    },
+    {
+        name: 'Google Analytics',
+        slug: 'analytics',
+        description: 'Want to know how your website is doing? We can provide custom event tracking for your website, generate regular customised reports and make recommendations.',
+        price: '',
+    },
+    {
+        name: 'HTTPS & SSL',
+        slug: 'https',
+        description: "Security is important. We can upgrade your website to be served over HTTPS (the 's' stands for secure).",
+        price: '',
+    },
+];
+
+export default services;
