+++ src/scss/layouts/home.scss
@@ -38,7 +38,7 @@ section.parallax {
         }
 
         &.slide-1 {
-            background-image: url('/img/parallax/unsplash_400_266.jpg');
+            background-image: url('../../img/parallax/unsplash_400_266.jpg');
         }
 
     }
@@ -109,7 +109,7 @@ body {
     section.parallax {
         .bcg {
             &.slide-1 {
-                background-image: url('/img/parallax/unsplash_400_266.jpg');
+                background-image: url('../../img/parallax/unsplash_400_266.jpg');
             }
         }
 
@@ -125,7 +125,7 @@ body {
     section.parallax {
         .bcg {
             &.slide-1 {
-                background-image: url('/img/parallax/unsplash_600_399.jpg');
+                background-image: url('../../img/parallax/unsplash_600_399.jpg');
             }
         }
     }
@@ -135,7 +135,7 @@ body {
     section.parallax {
         .bcg {
             &.slide-1 {
-                background-image: url('/img/parallax/unsplash_800_532.jpg');
+                background-image: url('../../img/parallax/unsplash_800_532.jpg');
             }
         }
     }
@@ -145,7 +145,7 @@ body {
     section.parallax {
         .bcg {
             &.slide-1 {
-                background-image: url('/img/parallax/unsplash_1000_665.jpg');
+                background-image: url('../../img/parallax/unsplash_1000_665.jpg');
             }
         }
     }
@@ -155,7 +155,7 @@ body {
     section.parallax {
         .bcg {
             &.slide-1 {
-                background-image: url('/img/parallax/unsplash_1200_798.jpg');
+                background-image: url('../../img/parallax/unsplash_1200_798.jpg');
             }
         }
     }
@@ -165,7 +165,7 @@ body {
     section.parallax {
         .bcg {
             &.slide-1 {
-                background-image: url('/img/parallax/unsplash_2600_1729.jpg');
+                background-image: url('../../img/parallax/unsplash_2600_1729.jpg');
             }
         }
     }
